Adding basic image load test cases
